requirements.txt
# Core Framework
flask==3.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0

# Jira Integration
jira==3.5.2
requests==2.31.0

# OpenAI
openai==1.3.7

# Email
sendgrid==6.10.0

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0
sentry-sdk==1.38.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2023.3

# PDF Generation (for Google Docs export)
reportlab==4.0.7

# Utilities
click==8.1.7
